refactor(setup): split StepScan into step-scan-model.ts, SourceSummary.tsx, and StepScan orchestration (CU-20)
649L → 258L StepScan; scanAndClassify + queryPhase + RenderTarget in model; #513 reconciliation logic stays with SourceSummary.
