+++ apps/desktop/src/features/setup/steps/StepScan.tsx
@@ -7,47 +7,23 @@
 // inbox_classify per source, showing per-source progress and a detection
 // summary.  Approval stays in the Inbox — no inbox_confirm called here.
 
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo } from 'react';
 import { useQueries, useQuery } from '@tanstack/react-query';
-import { Pill } from '@/ui/Pill';
-import type { PillVariant } from '@/ui/Pill';
-import { Table } from '@/ui';
-import type { TableColumn, TableRow } from '@/ui';
 import { m } from '@/lib/i18n';
-import { frameTypeCountLabel, masterLabel } from '@/lib/master-label';
 import { commands } from '@/bindings/index';
 import { unwrap } from '@/api/ipc';
-import type { InboxItemSummary } from '@/bindings/index';
-import type { InboxClassifyResponse } from '@/bindings/aliases';
 import type { SourceEntry } from '../sources-store';
 import type { FlushResult } from '../sources-store';
-import { sourceKindLabel } from '../sources-store';
 import { errMessage } from '@/lib/errors';
-import { RootDetectionConfig } from '@/features/inventory/RootDetectionConfig';
+import {
+  isScanTarget,
+  queryPhase,
+  scanAndClassify,
+  type RenderTarget,
+} from './step-scan-model';
+import { SourceSummary, type SourceScanState } from './SourceSummary';
 
-// ── Types ─────────────────────────────────────────────────────────────────────
-
-type ScanPhase = 'pending' | 'scanning' | 'done' | 'error';
-
-interface SourceScanState {
-  source: SourceEntry;
-  /** rootId returned by roots_register_batch for this source. */
-  rootId: string;
-  phase: ScanPhase;
-  items: InboxItemSummary[];
-  /** Per-item classify results keyed by inboxItemId. */
-  classifications: Map<string, InboxClassifyResponse>;
-  /**
-   * Items whose classify call FAILED, as opposed to succeeding with nothing to
-   * report. Without this the two are indistinguishable downstream: a missing
-   * map entry renders exactly like an empty classification, so a crashed
-   * folder showed an em dash and silently dropped out of the per-kind chips
-   * while still counting toward the header's file total — the same
-   * reconciliation gap issue #513 closed for unmapped IMAGETYP.
-   */
-  classifyFailures: Set<string>;
-  error?: string;
-}
+// ── Props ─────────────────────────────────────────────────────────────────────
 
 export interface StepScanProps {
   /** All registered sources from the wizard state. */
@@ -63,32 +39,6 @@ export interface StepScanProps {
 
 // ── Helpers ───────────────────────────────────────────────────────────────────
 
-function phaseLabel(phase: ScanPhase): string {
-  switch (phase) {
-    case 'pending':
-      return m.setup_scan_phase_pending();
-    case 'scanning':
-      return m.setup_scan_scanning();
-    case 'done':
-      return m.setup_scan_phase_done();
-    case 'error':
-      return m.settings_advanced_log_error();
-  }
-}
-
-function phasePillVariant(phase: ScanPhase): PillVariant {
-  switch (phase) {
-    case 'pending':
-      return 'neutral';
-    case 'scanning':
-      return 'warn';
-    case 'done':
-      return 'ok';
-    case 'error':
-      return 'danger';
-  }
-}
-
 /** Derive a stable rootId string from the flushResult for a given path.
  *  Falls back to the path itself when no root was returned (e.g. registration
  *  failed — we still try to scan using the path directly). */
@@ -96,347 +46,6 @@ function getRootId(flushResult: FlushResult, path: string): string {
   const row = flushResult.results.find((r) => r.path === path);
   // Successful rows carry the assigned rootId; fall back to the path if absent.
   return row?.rootId ?? path;
-}
-
-function scanKindCountLabel(kind: string, count: number): string {
-  switch (kind) {
-    case 'master':
-      return m.setup_scan_master_count({ count });
-    case 'unclassified':
-      return m.setup_scan_unclassified_count({ count });
-    case 'failed':
-      return m.setup_scan_classify_failed_count({ count });
-    default:
-      return frameTypeCountLabel(kind, count);
-  }
-}
-
-// ── Per-source detection summary ──────────────────────────────────────────────
-
-interface SourceSummaryProps {
-  state: SourceScanState;
-}
-
-function SourceSummary({ state }: SourceSummaryProps) {
-  const {
-    source,
-    rootId,
-    phase,
-    items,
-    classifications,
-    classifyFailures,
-    error,
-  } = state;
-  const [expanded, setExpanded] = useState(false);
-
-  // Spec 048 US4 T035: the wizard's Scan step is the earliest point a newly
-  // added raw/calibration root has a real backend id (roots.register.batch
-  // already ran via flushToDB before this step mounts) — a dedicated new
-  // wizard step was rejected as unwarranted churn for one control (see
-  // RootDetectionConfig's doc comment). getRootId() falls back to the raw
-  // path string when registration failed; skip rendering in that case since
-  // it is not a valid root id.
-  const hasRegisteredRoot = rootId !== source.path;
-  const isFrameSource =
-    source.kind === 'light_frames' || source.kind === 'calibration';
-
-  const totalItems = items.length;
-  const totalFiles = items.reduce((acc, it) => acc + it.fileCount, 0);
-
-  // Aggregate breakdown counts across all items, reconciled against the file
-  // count (issue #513): masters and unclassified files were previously
-  // dropped from this summary because they don't appear in a classify
-  // breakdown entry, so the chip counts silently undercounted totalFiles.
-  const kindCounts = new Map<string, number>();
-  for (const item of items) {
-    if (item.isMaster) {
-      kindCounts.set(
-        'master',
-        (kindCounts.get('master') ?? 0) + item.fileCount,
-      );
-      continue;
-    }
-    if (classifyFailures.has(item.inboxItemId)) {
-      // Count them so the chips still add up to the header total; a folder that
-      // failed must not simply vanish from the breakdown.
-      kindCounts.set(
-        'failed',
-        (kindCounts.get('failed') ?? 0) + item.fileCount,
-      );
-      continue;
-    }
-    const cls = classifications.get(item.inboxItemId);
-    for (const entry of cls?.breakdown ?? []) {
-      kindCounts.set(
-        entry.kind,
-        (kindCounts.get(entry.kind) ?? 0) + entry.count,
-      );
-    }
-    const unclassifiedCount = cls?.unclassifiedFiles.length ?? 0;
-    if (unclassifiedCount > 0) {
-      kindCounts.set(
-        'unclassified',
-        (kindCounts.get('unclassified') ?? 0) + unclassifiedCount,
-      );
-    }
-  }
-
-  // Stable sort for the table: by lane, then folder path.
-  const sortedItems = [...items].sort(
-    (a, b) =>
-      a.lane.localeCompare(b.lane) ||
-      a.relativePath.localeCompare(b.relativePath),
-  );
-
-  // Built per render rather than hoisted to module scope: the `m.*` accessors
-  // resolve against the active locale at call time, so hoisting would freeze
-  // the header labels to whichever locale was active at import.
-  const columns: TableColumn[] = [
-    { key: 'folder', label: m.setup_scan_col_folder() },
-    { key: 'files', label: m.setup_scan_col_files() },
-    { key: 'format', label: m.setup_scan_col_format() },
-    { key: 'types', label: m.setup_scan_col_types() },
-  ];
-
-  const rows: TableRow[] = sortedItems.map((item) => {
-    const cls = classifications.get(item.inboxItemId);
-    const typeParts = (cls?.breakdown ?? []).map((b) =>
-      frameTypeCountLabel(b.kind, b.count),
-    );
-    // Reconcile against fileCount (issue #513): files with no IMAGETYP (or an
-    // unmapped one) don't appear in the classify breakdown at all, so the type
-    // list previously undercounted the row's file total with no indication why.
-    const unclassifiedCount = cls?.unclassifiedFiles.length ?? 0;
-    if (unclassifiedCount > 0) {
-      typeParts.push(
-        m.setup_scan_unclassified_count({ count: unclassifiedCount }),
-      );
-    }
-    const types = typeParts.join(', ');
-    return {
-      _testid: `scan-item-${item.inboxItemId}`,
-      folder: (
-        <span className="pv-table__cell-inline">
-          {item.isMaster && <Pill variant="info">{m.setup_scan_master()}</Pill>}
-          {/* The root/top-level row of an expanded folder carries an empty
-              relativePath; label it instead of leaving it blank (issue #513). */}
-          {item.relativePath || m.setup_scan_root_label()}
-        </span>
-      ),
-      files: item.fileCount,
-      format: (item.format ?? item.lane).toUpperCase(),
-      // Individual calibration masters carry their frame type on the item
-      // itself (spec 040 FR-006); grouped sub-frame folders rely on the
-      // classify breakdown instead.
-      types: item.isMaster
-        ? masterLabel(item)
-        : classifyFailures.has(item.inboxItemId)
-          ? m.setup_scan_classify_failed()
-          : types || '—',
-    };
-  });
-
-  // The detail panel (chips + table) is expandable only when scan is done and
-  // there are items to show.
-  const isExpandable = phase === 'done' && totalItems > 0;
-
-  // Compact count summary for the always-visible header line.
-  const countSummary =
-    phase === 'done' && totalItems > 0
-      ? `${m.setup_scan_folder_count({ count: totalItems })} · ${m.setup_scan_file_count({ count: totalFiles })}`
-      : null;
-
-  return (
-    <div
-      data-testid={`scan-source-${source.path}`}
-      className="pv-setup-scan__card"
-    >
-      {/* ── Always-visible header row ─────────────────────────────────────── */}
-      {/* eslint-disable-next-line jsx-a11y/no-static-element-interactions -- conditionally interactive: role=button + keyboard (Enter/Space) handler applied together only when isExpandable; inert div otherwise */}
-      <div
-        role={isExpandable ? 'button' : undefined}
-        // eslint-disable-next-line jsx-a11y/no-noninteractive-tabindex -- tabIndex is paired with role=button (above) only when isExpandable, making the element a focusable button
-        tabIndex={isExpandable ? 0 : undefined}
-        aria-expanded={isExpandable ? expanded : undefined}
-        onClick={isExpandable ? () => setExpanded((v) => !v) : undefined}
-        onKeyDown={
-          isExpandable
-            ? (e) => {
-                if (e.key === 'Enter' || e.key === ' ') {
-                  e.preventDefault();
-                  setExpanded((v) => !v);
-                }
-              }
-            : undefined
-        }
-        className={
-          'pv-setup-scan__header' +
-          (isExpandable ? ' pv-setup-scan__header--expandable' : '')
-        }
-      >
-        {/* Chevron — only shown when expandable */}
-        {isExpandable && (
-          <span aria-hidden="true" className="pv-setup-scan__chevron">
-            {expanded ? '▾' : '▸'}
-          </span>
-        )}
-
-        {/* Source path */}
-        <span className="pv-setup-scan__path">{source.path}</span>
-
-        {/* Kind label (muted, small) */}
-        <span className="pv-setup-scan__kind">
-          {sourceKindLabel(source.kind)}
-        </span>
-
-        {/* Compact count summary */}
-        {countSummary && (
-          <span className="pv-setup-scan__count">{countSummary}</span>
-        )}
-
-        {/* Phase pill */}
-        <span aria-hidden="true">
-          <Pill variant={phasePillVariant(phase)}>{phaseLabel(phase)}</Pill>
-        </span>
-      </div>
-
-      {phase !== 'error' && (
-        <span
-          className="pv-visually-hidden"
-          role="status"
-          aria-live="polite"
-          aria-atomic="true"
-        >
-          {m.setup_scan_source_status({
-            path: source.path,
-            status: phaseLabel(phase),
-          })}
-        </span>
-      )}
-
-      {/* Spec 048 US4 T035: per-root reconcile mode + detection triggers,
-          set up at add-time with documented defaults pre-selected. */}
-      {isFrameSource && hasRegisteredRoot && (
-        <RootDetectionConfig rootId={rootId} />
-      )}
-
-      {/* ── Always-visible transient states (below header, outside collapse) ── */}
-      {/* These are small/short and don't benefit from collapse. */}
-      {phase === 'error' && error && (
-        <p
-          className="pv-setup-scan__msg pv-setup-scan__msg--error"
-          role="alert"
-        >
-          {m.setup_scan_source_error({ path: source.path, error })}
-        </p>
-      )}
-
-      {phase === 'done' && totalItems === 0 && (
-        <p
-          data-testid="scan-empty"
-          className="pv-setup-scan__msg pv-setup-scan__msg--empty"
-        >
-          {m.setup_scan_nothing_detected()}
-        </p>
-      )}
-
-      {/* ── Collapsible detail panel (chips + table) ─────────────────────── */}
-      {isExpandable && expanded && (
-        <div className="pv-setup-scan__detail">
-          {/* Kind-count chips */}
-          {kindCounts.size > 0 && (
-            <div className="pv-setup-scan__chips">
-              {Array.from(kindCounts.entries()).map(([kind, count]) => (
-                <span key={kind} className="pv-setup-scan__chip">
-                  {scanKindCountLabel(kind, count)}
-                </span>
-              ))}
-            </div>
-          )}
-
-          {/* Scrollable metadata table of detected ingestion groups.
-              `virtualized` is used for its scroll container + sticky header
-              (primitives.css `.pv-table__scroll`), not for windowing — a
-              per-source detection list is small. */}
-          <Table
-            columns={columns}
-            rows={rows}
-            virtualized
-            scrollClassName="pv-setup-scan__table-wrap"
-          />
-        </div>
-      )}
-    </div>
-  );
-}
-
-// One scan+classify pass for a single source. Classify failures are captured
-// per-item (never thrown) so one bad folder doesn't fail the whole query —
-// only a `scanFolder` failure itself surfaces as the query's error.
-async function scanAndClassify(
-  rootId: string,
-  rootAbsolutePath: string,
-): Promise<{
-  items: InboxItemSummary[];
-  classifications: Map<string, InboxClassifyResponse>;
-  classifyFailures: Set<string>;
-}> {
-  const scanResponse = unwrap(
-    await commands.inboxScanFolder({ rootId, rootAbsolutePath }),
-  );
-  const items = scanResponse.items ?? [];
-
-  const classifications = new Map<string, InboxClassifyResponse>();
-  const classifyFailures = new Set<string>();
-  await Promise.allSettled(
-    items.map(async (item) => {
-      try {
-        const cls = unwrap(
-          await commands.inboxClassify({
-            inboxItemId: item.inboxItemId,
-            rootAbsolutePath,
-          }),
-        );
-        classifications.set(item.inboxItemId, cls);
-      } catch {
-        // Don't abort the whole scan for one item — but do record it.
-        // Swallowing it entirely made a crash look like a clean "nothing
-        // detected" result.
-        classifyFailures.add(item.inboxItemId);
-      }
-    }),
-  );
-  return { items, classifications, classifyFailures };
-}
-
-/** A source resolved to a real backend root, ready to scan+classify. */
-interface ScanRenderTarget {
-  source: SourceEntry;
-  rootId: string;
-  kind: 'scan';
-}
-/** A source whose already-scanned status couldn't be resolved (roots.list
- * failed or returned no match) — surfaced as a static error, never scanned. */
-interface ResolutionErrorRenderTarget {
-  source: SourceEntry;
-  rootId: string;
-  kind: 'resolution-error';
-}
-type RenderTarget = ScanRenderTarget | ResolutionErrorRenderTarget;
-
-function isScanTarget(t: RenderTarget): t is ScanRenderTarget {
-  return t.kind === 'scan';
-}
-
-/** Map a `useQueries` result onto the same phase vocabulary the (still
- * hand-rolled) `pending`/`scanning` display states used. */
-function queryPhase(q: {
-  status: 'pending' | 'error' | 'success';
-  fetchStatus: 'fetching' | 'paused' | 'idle';
-}): ScanPhase {
-  if (q.status === 'success') return 'done';
-  if (q.status === 'error') return 'error';
-  return q.fetchStatus === 'fetching' ? 'scanning' : 'pending';
 }
 
 // ── StepScan ─────────────────────────────────────────────────────────────────
